telegram_bot/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorlog==6.7.0
et-xmlfile==1.1.0
Flask==3.0.0
frozenlist==1.4.0
greenlet==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openpyxl==3.1.2
psycopg2==2.9.9
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==3.0.0
yarl==1.9.2
